+++ app/[locale]/api/auth/[...nextauth]/route.ts
@@ -5,6 +5,7 @@
 
 // export { handler as GET, handler as POST }
 
+import { authConfig } from '@/configs/auth'
 import { Backend_URL } from '@/lib/Constants'
 
 import { NextAuthOptions } from 'next-auth'
@@ -12,79 +13,6 @@ import { JWT } from 'next-auth/jwt'
 import NextAuth from 'next-auth/next'
 import CredentialsProvider from 'next-auth/providers/credentials'
 
-async function refreshToken(token: JWT): Promise<JWT> {
-    const res = await fetch(Backend_URL + '/auth/refresh', {
-        method: 'POST',
-        headers: {
-            authorization: `Refresh ${token.backendTokens.refreshToken}`,
-        },
-    })
-    console.log('refreshed')
-
-    const response = await res.json()
-
-    return {
-        ...token,
-        backendTokens: response,
-    }
-}
-
-export const authOptions: NextAuthOptions = {
-    providers: [
-        CredentialsProvider({
-            name: 'Credentials',
-            credentials: {
-                username: {
-                    label: 'Username',
-                    type: 'text',
-                    placeholder: 'jsmith',
-                },
-                password: { label: 'Password', type: 'password' },
-            },
-            async authorize(credentials, req) {
-                if (!credentials?.username || !credentials?.password)
-                    return null
-                const { username, password } = credentials
-                const res = await fetch(Backend_URL + '/auth/signin', {
-                    method: 'POST',
-                    body: JSON.stringify({
-                        username,
-                        password,
-                    }),
-                    headers: {
-                        'Content-Type': 'application/json',
-                    },
-                })
-                if (res.status == 401) {
-                    console.log(res.statusText)
-
-                    return null
-                }
-                const user = await res.json()
-                return user
-            },
-        }),
-    ],
-
-    callbacks: {
-        async jwt({ token, user }) {
-            if (user) return { ...token, ...user }
-
-            if (new Date().getTime() < token.backendTokens.expiresIn)
-                return token
-
-            return await refreshToken(token)
-        },
-
-        async session({ token, session }) {
-            session.user = token.user
-            session.backendTokens = token.backendTokens
-
-            return session
-        },
-    },
-}
-
-const handler = NextAuth(authOptions)
+const handler = NextAuth(authConfig)
 
 export { handler as GET, handler as POST }
